chore: move translate_zones into new utility module
