+++ prereise/utility/tests/test_translate_zones.py
@@ -3,7 +3,7 @@ import pandas as pd
 import pytest
 from shapely.geometry import Polygon
 
-from prereise.gather.translate_zones import (
+from prereise.utility.translate_zones import (
     filter_interesting_zones,
     format_zone_df,
     make_zones_valid,
